# Support encoding json.Number type