--- a/cmd/jsontoml/main.go
+++ b/cmd/jsontoml/main.go
@@ -19,6 +19,7 @@ package main
 
 import (
 	"encoding/json"
+	"flag"
 	"io"
 
 	"github.com/pelletier/go-toml/v2"
@@ -33,6 +34,10 @@ Reading from a file:
   jsontoml file.json > file.toml
 `
 
+var (
+	useNumber = flag.Bool("use-number", false, "Tells the json decoder to unmarshal numbers into json.Number type instead of float64")
+)
+
 func main() {
 	p := cli.Program{
 		Usage: usage,
@@ -45,11 +50,17 @@ func convert(r io.Reader, w io.Writer) error {
 	var v interface{}
 
 	d := json.NewDecoder(r)
+	e := toml.NewEncoder(w)
+
+	if useNumber != nil && *useNumber {
+		d.UseNumber()
+		e.SetJsonNumber(true)
+	}
+
 	err := d.Decode(&v)
 	if err != nil {
 		return err
 	}
 
-	e := toml.NewEncoder(w)
 	return e.Encode(v)
 }