--- a/cmd/jsontoml/main_test.go
+++ b/cmd/jsontoml/main_test.go
@@ -25,7 +25,7 @@ func TestConvert(t *testing.T) {
   }
 }`,
 			expected: `[mytoml]
-a = 42.0
+a = 42
 `,
 		},
 		{