--- a/marshaler.go
+++ b/marshaler.go
@@ -3,6 +3,7 @@ package toml
 import (
 	"bytes"
 	"encoding"
+	"encoding/json"
 	"fmt"
 	"io"
 	"math"
@@ -252,6 +253,16 @@ func (enc *Encoder) encode(b []byte, ctx encoderCtx, v reflect.Value) ([]byte, e
 		return append(b, x.String()...), nil
 	case LocalDateTime:
 		return append(b, x.String()...), nil
+	case json.Number:
+		if x == "" { /// Useful zero value.
+			return append(b, "0"...), nil
+		} else if v, err := x.Int64(); err == nil {
+			return enc.encode(b, ctx, reflect.ValueOf(v))
+		} else if f, err := x.Float64(); err == nil {
+			return enc.encode(b, ctx, reflect.ValueOf(f))
+		} else {
+			return nil, fmt.Errorf("toml: unable to convert %q to int64 or float64", x)
+		}
 	}
 
 	hasTextMarshaler := v.Type().Implements(textMarshalerType)