--- a/marshaler_test.go
+++ b/marshaler_test.go
@@ -682,31 +682,6 @@ L = 2.2
 `,
 		},
 		{
-			desc: "json numbers",
-			v: struct {
-				A json.Number
-				B json.Number
-				C json.Number
-				D json.Number
-				E json.Number
-				F json.Number
-			}{
-				A: "1.1",
-				B: "42e-3",
-				C: "42",
-				D: "0",
-				E: "0.0",
-				F: "",
-			},
-			expected: `A = 1.1
-B = 0.042
-C = 42
-D = 0
-E = 0.0
-F = 0
-`,
-		},
-		{
 			desc: "comments",
 			v: struct {
 				Table comments `comment:"Before table"`
@@ -969,6 +944,29 @@ func TestEncoderSetIndentSymbol(t *testing.T) {
 	require.NoError(t, err)
 	expected := `[parent]
 >>>hello = 'world'
+`
+	assert.Equal(t, expected, w.String())
+}
+
+func TestEncoderSetJsonNumber(t *testing.T) {
+	var w strings.Builder
+	enc := toml.NewEncoder(&w)
+	enc.SetJsonNumber(true)
+	err := enc.Encode(map[string]interface{}{
+		"A": json.Number("1.1"),
+		"B": json.Number("42e-3"),
+		"C": json.Number("42"),
+		"D": json.Number("0"),
+		"E": json.Number("0.0"),
+		"F": json.Number(""),
+	})
+	require.NoError(t, err)
+	expected := `A = 1.1
+B = 0.042
+C = 42
+D = 0
+E = 0.0
+F = 0
 `
 	assert.Equal(t, expected, w.String())
 }