Adds encoding support for [`json.Number`](https://pkg.go.dev/encoding/json#Number) and fixes https://github.com/pelletier/go-toml/issues/897

On a side note, I am currently porting a codebase that uses https://github.com/BurntSushi/toml and the lack of encoding support for `json.Number` type in go-toml neat library was a blocker.

---

Paste `benchstat` results here
```
✦ ❯ ./ci.sh benchmark -d v2
Executing benchmark for v2 at /var/folders/yw/xfw4875j6xv69_2r5v52b1yc0000gn/T/tmp.gbbPAaT5wG
Preparing worktree (checking out 'v2')
HEAD is now at 34765b4 Fix unmarshaling of nested non-exported struct (#917)
/var/folders/yw/xfw4875j6xv69_2r5v52b1yc0000gn/T/tmp.gbbPAaT5wG ~/src/go-toml
PASS
ok      github.com/pelletier/go-toml/v2 0.524s
goos: darwin
goarch: arm64
pkg: github.com/pelletier/go-toml/v2/benchmark
BenchmarkUnmarshalDataset/config-2                    97          11903903 ns/op          88.10 MB/s     5769411 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          11925088 ns/op          87.94 MB/s     5769267 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          11915622 ns/op          88.01 MB/s     5769277 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    99          11900315 ns/op          88.12 MB/s     5769391 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                    99          11971822 ns/op          87.60 MB/s     5769346 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          12181822 ns/op          86.09 MB/s     5769320 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          12227300 ns/op          85.77 MB/s     5769267 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          12061128 ns/op          86.95 MB/s     5769357 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          12063997 ns/op          86.93 MB/s     5769406 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/config-2                   100          12059835 ns/op          86.96 MB/s     5769332 B/op     219177 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45043838 ns/op          48.87 MB/s    80013609 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    26          45198702 ns/op          48.70 MB/s    80013609 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          44814930 ns/op          49.12 MB/s    80013610 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          44953443 ns/op          48.97 MB/s    80013606 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    26          45302455 ns/op          48.59 MB/s    80013612 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45037880 ns/op          48.88 MB/s    80013607 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    26          44884133 ns/op          49.05 MB/s    80013612 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45058027 ns/op          48.86 MB/s    80013608 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    25          45323782 ns/op          48.57 MB/s    80013611 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/canada-2                    26          45093024 ns/op          48.82 MB/s    80013612 B/op     670398 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12663785 ns/op          44.07 MB/s    34501828 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              90          13116770 ns/op          42.54 MB/s    34501832 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12798680 ns/op          43.60 MB/s    34501754 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12771673 ns/op          43.69 MB/s    34501898 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              88          13127312 ns/op          42.51 MB/s    34501745 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              88          12969035 ns/op          43.03 MB/s    34501803 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              91          12884934 ns/op          43.31 MB/s    34501748 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              87          12880178 ns/op          43.33 MB/s    34501914 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              85          13174064 ns/op          42.36 MB/s    34501830 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/citm_catalog-2              88          13031836 ns/op          42.82 MB/s    34501887 B/op     177994 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5508132 ns/op          80.22 MB/s    12633326 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  216           5550444 ns/op          79.61 MB/s    12633192 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  208           5429029 ns/op          81.39 MB/s    12633339 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5415060 ns/op          81.60 MB/s    12633134 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5417146 ns/op          81.57 MB/s    12633278 B/op      54756 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  220           5431102 ns/op          81.36 MB/s    12633779 B/op      54757 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5404494 ns/op          81.76 MB/s    12632923 B/op      54754 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5397776 ns/op          81.86 MB/s    12632811 B/op      54754 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5402685 ns/op          81.79 MB/s    12633137 B/op      54755 allocs/op
BenchmarkUnmarshalDataset/twitter-2                  222           5437698 ns/op          81.26 MB/s    12633566 B/op      54757 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          57018850 ns/op          47.07 MB/s    21272713 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56443510 ns/op          47.55 MB/s    21272708 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56607523 ns/op          47.41 MB/s    21272716 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56613048 ns/op          47.41 MB/s    21272720 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56461679 ns/op          47.54 MB/s    21272715 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56510048 ns/op          47.50 MB/s    21272710 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56987342 ns/op          47.10 MB/s    21272711 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56552406 ns/op          47.46 MB/s    21272713 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56843088 ns/op          47.22 MB/s    21272720 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/code-2                      20          56659192 ns/op          47.37 MB/s    21272719 B/op    1000531 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101501 ns/op          79.80 MB/s      185113 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101348 ns/op          79.92 MB/s      185110 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101601 ns/op          79.72 MB/s      185111 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101342 ns/op          79.93 MB/s      185107 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101900 ns/op          79.49 MB/s      185105 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101281 ns/op          79.98 MB/s      185109 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            101227 ns/op          80.02 MB/s      185110 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            103397 ns/op          78.34 MB/s      185118 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102660 ns/op          78.90 MB/s      185107 B/op       1305 allocs/op
BenchmarkUnmarshalDataset/example-2                10000            102328 ns/op          79.16 MB/s      185121 B/op       1305 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3697077               325.0 ns/op        33.85 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3695973               330.6 ns/op        33.27 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3688902               325.9 ns/op        33.76 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3705112               324.0 ns/op        33.95 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3692125               324.2 ns/op        33.93 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3703725               325.0 ns/op        33.84 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3703592               323.8 ns/op        33.97 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3709260               323.6 ns/op        33.99 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3695694               324.6 ns/op        33.88 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/struct-2       3705052               323.9 ns/op        33.96 MB/s         805 B/op          9 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2617071               454.8 ns/op        24.19 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2633796               457.0 ns/op        24.07 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2623290               457.2 ns/op        24.06 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2622032               458.0 ns/op        24.02 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2633118               456.5 ns/op        24.09 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2621324               457.5 ns/op        24.04 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2641050               456.1 ns/op        24.12 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2619297               456.8 ns/op        24.08 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2629777               455.4 ns/op        24.15 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/SimpleDocument/map-2          2624958               456.1 ns/op        24.12 MB/s        1133 B/op         13 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41130             29183 ns/op         179.59 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41002             29185 ns/op         179.58 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41238             29131 ns/op         179.91 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41224             29133 ns/op         179.90 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41338             29146 ns/op         179.82 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41221             29092 ns/op         180.15 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41212             29178 ns/op         179.62 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41218             29172 ns/op         179.66 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          40947             29245 ns/op         179.21 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/struct-2          41004             29542 ns/op         177.41 MB/s       20466 B/op        160 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27697             43163 ns/op         121.42 MB/s       37746 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27596             43183 ns/op         121.37 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27771             43180 ns/op         121.38 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27648             43180 ns/op         121.38 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27590             43176 ns/op         121.39 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27776             43212 ns/op         121.28 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27733             43269 ns/op         121.13 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27636             43142 ns/op         121.48 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27720             43237 ns/op         121.22 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/ReferenceFile/map-2             27654             43214 ns/op         121.28 MB/s       37747 B/op        619 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              162596              7305 ns/op          74.75 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              162312              7260 ns/op          75.20 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              163777              7265 ns/op          75.16 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              163282              7261 ns/op          75.19 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              162428              7261 ns/op          75.20 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              162064              7247 ns/op          75.34 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              162523              7276 ns/op          75.04 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              163615              7350 ns/op          74.29 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              160548              7469 ns/op          73.10 MB/s        7441 B/op        141 allocs/op
BenchmarkUnmarshal/HugoFrontMatter-2              161103              7353 ns/op          74.25 MB/s        7441 B/op        141 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4492448               264.3 ns/op        45.41 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4613572               264.7 ns/op        45.34 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4595530               262.6 ns/op        45.69 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4612095               265.8 ns/op        45.14 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4556260               263.8 ns/op        45.49 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4620241               260.1 ns/op        46.14 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4616716               259.3 ns/op        46.28 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4592475               259.1 ns/op        46.31 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4607875               259.6 ns/op        46.22 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/struct-2         4634221               260.0 ns/op        46.15 MB/s         240 B/op          8 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3364288               355.7 ns/op        33.74 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3361501               356.3 ns/op        33.68 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3370617               356.3 ns/op        33.68 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3373314               356.0 ns/op        33.70 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3364156               356.9 ns/op        33.62 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3364855               356.7 ns/op        33.64 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3367714               357.3 ns/op        33.58 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3365958               356.1 ns/op        33.69 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3364312               356.0 ns/op        33.70 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/SimpleDocument/map-2            3360133               363.8 ns/op        32.98 MB/s         272 B/op          9 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54135             22101 ns/op          93.07 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54378             21997 ns/op          93.51 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54459             22032 ns/op          93.36 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54476             22058 ns/op          93.26 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54655             22077 ns/op          93.18 MB/s       28304 B/op        317 allocs/op
BenchmarkMarshal/ReferenceFile/struct-2            54340             22038 ns/op          93.34 MB/s       28304 B/op        317 allocs/op
…