@dangra 2024-01-24
@pelletier this PR looks ready for another pass and hopefully (squash) merge it into `v2` branch 🙏